requirements.txt
python-dotenv
fastapi
uvicorn
httpx
fal-client
